requirements.txt
biopython==1.70
click==6.7
numpy==1.13.3
